+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/interop/AnvilRuntimeTestServices.kt
@@ -31,7 +31,10 @@ class AnvilRuntimeEnvironmentConfigurator(testServices: TestServices) :
     configuration: CompilerConfiguration,
     module: TestModule,
   ) {
-    if (MetroDirectives.WITH_ANVIL in module.directives) {
+    if (
+      MetroDirectives.WITH_ANVIL in module.directives ||
+        MetroDirectives.ENABLE_ANVIL_KSP in module.directives
+    ) {
       for (file in anvilRuntimeClasspath) {
         configuration.addJvmClasspathRoot(file)
       }
@@ -48,7 +51,10 @@ class AnvilRuntimeClassPathProvider(testServices: TestServices) :
   RuntimeClasspathProvider(testServices) {
   override fun runtimeClassPaths(module: TestModule): List<File> {
     return buildList {
-      if (MetroDirectives.WITH_ANVIL in module.directives) {
+      if (
+        MetroDirectives.WITH_ANVIL in module.directives ||
+          MetroDirectives.ENABLE_ANVIL_KSP in module.directives
+      ) {
         addAll(anvilRuntimeClasspath)
       }
       if (MetroDirectives.WITH_KI_ANVIL in module.directives) {