+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/MetroSimpleFunction.kt
@@ -10,7 +10,9 @@ import org.jetbrains.kotlin.name.CallableId
 import org.jetbrains.kotlin.synthetic.isVisibleOutside
 
 /** Simple holder with resolved annotations to save us lookups. */
-// TODO cache these in a transformer context?
+// TODO
+//  cache these in a transformer context?
+//  make this MetroCallable to support properties
 @Poko
 internal class MetroSimpleFunction(
   @Poko.Skip val ir: IrSimpleFunction,
@@ -19,9 +21,8 @@ internal class MetroSimpleFunction(
 ) : Comparable<MetroSimpleFunction> {
   override fun toString() = callableId.toString()
 
-  override fun compareTo(other: MetroSimpleFunction): Int {
-    return callableId.toString().compareTo(other.callableId.toString())
-  }
+  override fun compareTo(other: MetroSimpleFunction): Int =
+    callableId.toString().compareTo(other.callableId.toString())
 }
 
 internal val MetroSimpleFunction.isAccessorCandidate: Boolean