+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/ir.kt
@@ -141,6 +141,7 @@ import org.jetbrains.kotlin.ir.util.getSimpleFunction
 import org.jetbrains.kotlin.ir.util.getValueArgument
 import org.jetbrains.kotlin.ir.util.hasAnnotation
 import org.jetbrains.kotlin.ir.util.hasShape
+import org.jetbrains.kotlin.ir.util.isFromJava
 import org.jetbrains.kotlin.ir.util.isObject
 import org.jetbrains.kotlin.ir.util.isStatic
 import org.jetbrains.kotlin.ir.util.kotlinFqName
@@ -213,7 +214,7 @@ internal fun IrType.rawType(): IrClass {
         when {
           this is IrErrorType -> "Error type encountered: ${dumpKotlinLike()}"
           classifierOrNull is IrTypeParameterSymbol ->
-            "Type parameter encountered: ${dumpKotlinLike()}"
+            "Unexpected type parameter encountered: ${dumpKotlinLike()}"
           else -> "Unrecognized type! ${dumpKotlinLike()} (${classifierOrNull?.javaClass})"
         }
       reportCompilerBug(message)
@@ -653,7 +654,10 @@ internal fun assignConstructorParamsToFields(
           )
           .apply {
             isFinal = true
-            initializer = context.createIrBuilder(symbol).run { irExprBody(irGet(irParameter.ir)) }
+            initializer =
+              context.createIrBuilder(symbol).run {
+                irExprBody(irGet(irParameter.asValueParameter))
+              }
           }
       put(irParameter, irField)
     }
@@ -1460,6 +1464,7 @@ private fun buildDeepSubstitutionMap(
 
 private class DeepTypeSubstitutor(private val substitutionMap: Map<IrTypeParameterSymbol, IrType>) :
   TypeRemapper {
+  private val mapByName = substitutionMap.mapKeys { it.key.owner.name.identifier }
   private val cache = mutableMapOf<IrType, IrType>()
 
   override fun enterScope(irTypeParametersContainer: IrTypeParametersContainer) {}
@@ -1754,3 +1759,11 @@ internal fun IrBuilderWithScope.irGetProperty(
 
 internal val IrConstructorCall.annotationClass: IrClass
   get() = symbol.owner.parentAsClass
+
+internal fun IrClass.companionObjectOrSelfIfObjectOrJava(): IrClass {
+  return when {
+    isFromJava() -> this
+    kind.isObject -> this
+    else -> companionObject() ?: this
+  }
+}